test: merge exporter overlay spec test into ExporterArgsOverlayTest
The former characterization test (now a spec test) belongs alongside the
unit tests for ExporterArgsOverlay. Moved into a ViaResolver nested class
so the MockEnvironment/Spring-Binder setup is scoped to that group only.
